Implement fixed power/pixel setting.
Using the `frequency' setting, a fixed pulsewidth is emitted every pixel.
This allows for sharp cuts regardless of cutting speed. By default it is set
to one second i.e. continuously on for current behaviour.

By setting the frequency to say 600, it emits a pulse of 1/600 second for
every pixel. This is independent from the power setting.